% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clara_clust}
\alias{clara_clust}
\title{CLARA clustering}
\usage{
clara_clust(
  data,
  nb_sample = 100,
  size_sample = 40 + 2 * nb_cluster,
  nb_cluster = 4,
  method = "LCS",
  plot = FALSE,
  find_best_method = "Distance",
  with.diss = TRUE,
  cores = 2
)
}
\arguments{
\item{data}{The dataset to use. In case of sequences, use seqdef (from TraMineR package) to create such an object.}

\item{nb_sample}{The number of subsets to test.}

\item{size_sample}{The size of each subset}

\item{nb_cluster}{The number of medoids}

\item{method}{The calculation method to compute the distance matrix. (See the function seqdist in TraMineR package)}

\item{plot}{Boolean variable to plot the result of clustering}

\item{find_best_method}{Method to select the best subset. "Distance" is for the mean distance and "DB" is for Davies-Bouldin value.}

\item{with.diss}{Boolean if the distance matrix should be returned}

\item{cores}{Number of cores to use for parallelism}
}
\value{
An object with the data, the medoids id (name of the line), the clustering and the distance matrix
}
\description{
With the help of TraMineR package, CLARA clustering provide a clustering of big dataset.\cr The main objective is to cluster state sequences with the "LCS" distance calculation method to find the best partition in N clusters.
}
\examples{

\dontrun{
library(TraMineR)
data(biofam)
#Basic CLARA Computing
my_cluster <- clara_clust(seqdef(biofam), method="LCS")

#Improved CLARA Computing
my_cluster <- clara_clust(seqdef(biofam), method="LCS", improve_method="Silhouette", nbessai_improved=30)
}
}
