% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clarans_clust}
\alias{clarans_clust}
\title{CLARANS clustering}
\usage{
clarans_clust(
  data,
  nb_cluster,
  distargs = list(method = "LCS"),
  maxneighbours,
  numlocal,
  plot = FALSE,
  cores = detectCores() - 1
)
}
\arguments{
\item{data}{The dataset to use. In case of sequences, use seqdef (from TraMineR package) to create such an object.}

\item{nb_cluster}{The number of medoids}

\item{distargs}{List with method parameters to apply. (See the function seqdist in TraMineR package)}

\item{maxneighbours}{Number of neighbours to explore to find a better clustering}

\item{numlocal}{Number of initialisation of the starting medoids}

\item{plot}{Boolean variable to plot the research convergence}

\item{cores}{Number of cores to use for parallelism}
}
\value{
An object with the data, the medoids id (name of the line), the clustering and the distance matrix
}
\description{
With the help of TraMineR package, CLARANS clustering provide a clustering of big dataset.\cr The main objective is to cluster state sequences with the "LCS" distance calculation method to find the best partition in N clusters. \cr \cr WARNING : this function is less efficient than cLARA.
}
\examples{

#creating sequences
library(TraMineR)
data(mvad)
mvad.labels <- c("employment", "further education", "higher education","joblessness", "school", "training")
mvad.scode <- c("EM", "FE", "HE", "JL", "SC", "TR")
mvad.seq <- seqdef(mvad, 17:86, states = mvad.scode,abels = mvad.labels, xtstep = 6)

#CLARANS Clustering
my_cluster <- clarans_clust(mvad.seq,nb_cluster = 4, maxneighbours = 20, numlocal = 4, plot = TRUE)
}
