% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{davies_bouldin}
\alias{davies_bouldin}
\title{Davies Bouldin Index}
\usage{
davies_bouldin(seq_obj, method, diss = TRUE, plot = TRUE, cores = 2)
}
\arguments{
\item{seq_obj}{The object generated with CLARA Clustering}

\item{method}{The calculation method to compute the distance matrix. (See the function seqdist in TraMineR package)}

\item{diss}{Boolean to express if the parameter diss from CLARA.seq clustering has been returned (Matrix size must be k columns and n rows - see refseq function from TraMineR package)}

\item{plot}{Boolean variable to plot the result}

\item{cores}{Number of cores to use for parallelism}
}
\value{
The value of the index
}
\description{
Implementation of Davies-Bouldin index to evaluate the quality of CLARA Clustering.
}
\examples{
\dontrun{
my_index <- db_index(my_cluster, "LCS", diss = FALSE)
}
}
