% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fuzzy_clust}
\alias{fuzzy_clust}
\title{FUZZY-CLARA clustering}
\usage{
fuzzy_clust(
  data,
  nb_sample = 100,
  size_sample = 40 + 2 * nb_cluster,
  nb_cluster = 4,
  method = "LCS",
  fuzzyfier = 2,
  p = 5,
  threshold = 10,
  max_iter = 10,
  noise = 0.5,
  plot = FALSE
)
}
\arguments{
\item{data}{The dataset to use. In case of sequences, use seqdef (from TraMineR package) to create such an object.}

\item{nb_sample}{The number of subsets to test.}

\item{size_sample}{The size of each subset}

\item{nb_cluster}{The number of medoids}

\item{method}{The calculation method to compute the distance matrix. (See the function seqdist in TraMineR package)}

\item{p}{Number of candidate to test to be a better medoid}

\item{threshold}{Variable to exclude outliers, whose values are greater than threshold}

\item{max_iter}{Number of maximal iteration to do to find the set of medoids}

\item{noise}{Small value to avoid divisions by 0 error}

\item{plot}{Boolean variable to plot the result of clustering}

\item{fuzzifier}{Value of the fuzzifier (default is 2, which is the traditionnal value)}

\item{cores}{Number of cores to use for parallelism}
}
\value{
An object with the data, the medoids id (name of the line), the clustering and the distance matrix
}
\description{
With the help of TraMineR package, FUZZY-CLARA clustering provide a clustering of big dataset.\cr The main objective is to cluster state sequences with the "LCS" distance calculation method to find the best partition in N clusters. \cr This function is a mix between CLARA and the ROBUST FUZZY C-MEDOIDS. \cr \cr WARNING : This function is not finished yet !
}
\examples{

\dontrun{
library(TraMineR)
data(biofam)
#Basic CLARA Computing
my_cluster <- fuzzy_clust(seqdef(biofam),nb_sample = 14, size_sample = 50, plot = TRUE, threshold = 7, max_iter = 10, p=5)
}
}
