% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.clara_seq}
\alias{plot.clara_seq}
\title{CLARA Plot Method}
\usage{
\method{plot}{clara_seq}(obj)
}
\arguments{
\item{obj}{An object from the class clara_seq (create by the function clara_clust())}
}
\value{
A plot of the mean distance's evolution
}
\description{
Create graph from CLARA object to show the evolution of the mean distance
}
\examples{
plot(my_cluster)
}
