% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.clarafuzzy_seq}
\alias{plot.clarafuzzy_seq}
\title{CLARA-FUZZY Plot Method}
\usage{
\method{plot}{clarafuzzy_seq}(obj)
}
\arguments{
\item{obj}{An object from the class clarafuzzy_seq (create by the function fuzzy_clust())}
}
\value{
A plot of the index evolution
}
\description{
Create graph from CLARA-FUZZY object to show the evolution of the index value
}
\examples{
plot(my_cluster)
}
