% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.clarans_seq}
\alias{plot.clarans_seq}
\title{CLARANS Plot Method}
\usage{
\method{plot}{clarans_seq}(obj)
}
\arguments{
\item{obj}{An object from the class clarans_seq (create by the function clarans_clust())}
}
\value{
A plot of the mean distance's evolution
}
\description{
Create graph from CLARANS object to show the evolution of the mean distance
}
\examples{
plot(my_cluster)
}
