% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clara_clust}
\alias{clara_clust}
\title{CLARA clustering}
\usage{
clara_clust(
  data,
  nb_sample = 100,
  size_sample = 40 + 2 * nb_cluster,
  nb_cluster = 4,
  distargs = list(method = "LCS"),
  plot = FALSE,
  find_best_method = "Distance",
  with.diss = TRUE,
  cores = detectCores() - 1
)
}
\arguments{
\item{data}{The dataset to use. In case of sequences, use seqdef (from TraMineR package) to create such an object.}

\item{nb_sample}{The number of subsets to test.}

\item{size_sample}{The size of each subset}

\item{nb_cluster}{The number of medoids}

\item{distargs}{List with method parameters to apply. (See the function seqdist in TraMineR package)}

\item{plot}{Boolean variable to plot the result of clustering}

\item{find_best_method}{Method to select the best subset. "Distance" is for the mean distance and "DB" is for Davies-Bouldin value.}

\item{with.diss}{Boolean if the distance matrix should be returned}

\item{cores}{Number of cores to use for parallelism}
}
\value{
An object of class clara_seq
}
\description{
With the help of TraMineR package, CLARA clustering provide a clustering of big dataset.\cr The main objective is to cluster state sequences with the "LCS" distance calculation method to find the best partition in N clusters.
}
\examples{

#creating sequences
library(TraMineR)
data(mvad)
mvad.labels <- c("employment", "further education", "higher education","joblessness", "school", "training")
mvad.scode <- c("EM", "FE", "HE", "JL", "SC", "TR")
mvad.seq <- seqdef(mvad, 17:86, states = mvad.scode,abels = mvad.labels, xtstep = 6)

#CLARA Clustering
my_cluster <- clara_clust(mvad.seq,nb_cluster = 4, nb_sample = 10, size_sample = 20, with.diss = TRUE)

#CLARA Clustering with Davies-Bouldin Method
my_cluster <- clara_clust(mvad.seq,nb_cluster = 4, nb_sample = 10, size_sample = 20, with.diss = TRUE, find_best_method = "DB")
}
