% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{davies_bouldin}
\alias{davies_bouldin}
\title{Davies Bouldin Index}
\usage{
davies_bouldin(
  seq_obj,
  distargs = list(method = "LCS"),
  diss = TRUE,
  plot = TRUE,
  cores = detectCores() - 1
)
}
\arguments{
\item{seq_obj}{The object generated with CLARA Clustering}

\item{distargs}{List with method parameters to apply. (See the function seqdist in TraMineR package)}

\item{diss}{Boolean to express if the parameter diss from CLARA.seq clustering has been returned (Matrix size must be k columns and n rows - see refseq function from TraMineR package)}

\item{plot}{Boolean variable to plot the research convergence}

\item{cores}{Number of cores to use for parallelism}
}
\value{
The value of the index
}
\description{
Implementation of Davies-Bouldin index to evaluate the quality of CLARA Clustering.
}
\examples{
\dontrun{
my_index <- davies_bouldin(my_cluster)
}
}
